Create README.md
Part 1

To approach this problem, I first downloaded the 'Used Bike Prices data in India' data set from Kaggle. The data was then cleaned using excel. After that, to define the problem, I edited the file and selected the columns within the data set that were useful for my scenario. This was done by analyzing the data set and looking for information that was useful. Upon examination, it was found that columns such as the Bike's model, Year, Kilometres driven, Owner, mileage, power, and price of the bikes were useless information as they are not applicable for providing a data-driven recommendation. Rather, the location of the seller is valuable as it indicates the city in India that the entrepreneur can open the new physical store in as well as the count/number of used bikes that were sold in that city. Therefore, to narrow my results and create a graph to depict this, I emphasized the "location" column in the new edited file called bike_edit.csv. Then, I used the value_counts()function to obtain a series containing counts of values of the bikes that were sold in each city in India. The results were arranged in descending order and plotted on a bar graph. This was done so that the first city as indicated by the x-axis on the plot represents the most-frequently occurring element which is the highest number of used bikes sold. This allowed me to determine the number of used bikes in each city in India, and then select the top 25 cities from that list to create a bar graph using map plot lib. As well, this allowed for a comparative analysis between the major cities in India and the corresponding used bikes that were sold at each location. Hence, to answer the question regarding which city in India the first physical store should be opened in, the count value for the number of bikes sold at each city location was used and measured. This was to determine which city location had the greatest sales, resulting in the most profits. Based on this data, this recommendation can provide valuable information to the entrepreneur regarding which city was able to sell the most used bikes. This means it is an optimal location to open a physical store as it can ensure a profitable business. Hence, as explained by the data analysis, the city with the highest count of used bikes sold will likely be the city that is most profitable for the entrepreneur’s financial success. After analyzing this, it was found that Delhi was the city that had the highest count of sales in used bikes in India.